+++ lib/sendfileportable.h
@@ -3,6 +3,6 @@
 #define _SENDFILEPORTABLE_H_
 #include <sys/types.h>
 
-int sendfileportable(int s, int fd, off_t offset, size_t nbytes, size_t *sbytes);
+int sendfileportable(int s, int fd, off_t offset, size_t nbytes, off_t *sbytes);
 
 #endif
